+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Version 1.5.1 (not released yet)
 - no longer require "-sound" argument when launching level editor with sound support
 - allow opening a level in the level editor by using `-level` command line argument
 - add option to associate `rfl` file extension with Dash Faction level editor in the setup program
+- fix level editor crash randomly occuring when opening cutscene properties
 
 Version 1.5.0 (released 2020-09-05)
 -----------------------------------